Add first integration test to validate plugin behavior on a representative OSM dataset

+++ lib/install.sh
@@ -18,6 +18,19 @@ if [ $? -ne 0 ]; then
 	exit 1
 fi
 
+# Install osmosis-core
+echo "Installing osmosis-xml library into the local Maven repository"
+mvn install:install-file \
+	-Dfile=osmosis-xml-0.41.jar \
+	-DgroupId=org.openstreetmap \
+	-DartifactId=osmosis-xml \
+	-Dversion=0.41 \
+	-Dpackaging=jar >> $OUTPUT
+if [ $? -ne 0 ]; then
+	echo "Failed: see $OUTPUT logfile for details"
+	exit 1
+fi
+
 # Clean
 echo "Done! Cleaning working files..."
 rm -f $OUTPUT
